+++ sdk/include/opentelemetry/sdk/configuration/composable_rule_based_sampler_rule_configuration.h
@@ -1,0 +1,19 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+#pragma once
+#include "opentelemetry/sdk/configuration/sampler_configuration.h"
+#include "opentelemetry/version.h"
+OPENTELEMETRY_BEGIN_NAMESPACE
+namespace sdk
+{
+namespace configuration
+{
+class ComposableRuleBasedSamplerRuleConfiguration
+{
+public:
+  ComposableRuleBasedSamplerRuleConfiguration()  = default;
+  ~ComposableRuleBasedSamplerRuleConfiguration() = default;
+};
+}  // namespace configuration
+}  // namespace sdk
+OPENTELEMETRY_END_NAMESPACE